+++ example/test.dart
@@ -2,4 +2,5 @@
 import 'package:happypass/happypass.dart';
 
 void main () async {
+
 }